+++ openpgl/directional/dqt/DQTSurfaceSamplingDistribution.h
@@ -39,6 +39,10 @@ struct DQTSurfaceSamplingDistribution: public ISurfaceSamplingDistribution {
         return distribution.samplePdf(sample, dir);
     }
 
+    inline float pdfLi(const Vector3 dir) const override {
+        return distribution.pdf(dir);
+    };
+
     inline bool validate() const override {
         return distribution.isValid();
     };