+++ openpgl/directional/vmm/VMMVolumeSamplingDistribution.h
@@ -122,6 +122,11 @@ struct __aligned(TVMMDistribution::VectorSize*4) VMMVolumeSamplingDistribution: 
         return pdf(dir);
     }
 
+    inline float pdfLi(const Vector3 dir) const override
+    {
+        return m_liDistribution.pdf(dir);
+    }
+
     inline bool validate() const override
     {
         return m_numDistributions > 0;